Changing the Manager
