+++ Package.swift
@@ -5,6 +5,7 @@ import PackageDescription
 
 let package = Package(
     name: "HumanSpeech",
+    platforms: [.iOS(.v15)],
     products: [
         // Products define the executables and libraries a package produces, making them visible to other packages.
         .library(
