--- Sources/HumanSpeech/HumanIdentifierManager.swift
@@ -1,67 +1,0 @@
-//
-//  HumanIdentifierManager.swift
-//  HumanSpeech
-//
-//  Created by Daniel Leal PImenta on 25/09/25.
-//
-
-import CoreML
-import SoundAnalysis
-import AVFoundation
-import SwiftUI
-import Speech
-
-
-@available(iOS 15.0, *)
-@MainActor
-class HumanIdentifierManager: NSObject, ObservableObject, SNResultsObserving {
-
-    private var analyzer: SNAudioStreamAnalyzer!
-    private let audioEngine = AVAudioEngine()
-    private var inputFormat: AVAudioFormat!
-
-    @Published var detectedSound: String = "Nenhum som detectado"
-
-    override init() {
-        super.init()
-
-        inputFormat = audioEngine.inputNode.inputFormat(forBus: 0)
-        analyzer = SNAudioStreamAnalyzer(format: inputFormat)
-
-        do {
-            let model = try HumanSpeaking(configuration: MLModelConfiguration())
-            let request = try SNClassifySoundRequest(mlModel: model.model)
-            try analyzer.add(request, withObserver: self)
-        } catch {
-            print("ERRO AO CARREGAR MODELO")
-        }
-    }
-
-    func start() {
-        audioEngine.inputNode.installTap(onBus: 0, bufferSize: 8192, format: inputFormat) {
-            buffer, time in
-            self.analyzer.analyze(buffer, atAudioFramePosition: time.sampleTime)
-        }
-
-        do {
-            try audioEngine.start()
-        } catch {
-            print("Erro ao iniciar áudio: (error)")
-        }
-    }
-
-    nonisolated func request(_ request: SNRequest, didProduce result: SNResult) {
-        if let classificationResult = result as? SNClassificationResult,
-           let classification = classificationResult.classifications.first {
-            let identifier = classification.identifier
-            DispatchQueue.main.async {
-                self.detectedSound = "\(identifier) ((String(classification.confidence * 100))%)"
-                print("SOM DETECTADO: (self.detectedSound)")
-            }
-        }
-    }
-}
-
-
-
-
